% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.pathways.R
\name{get.pathways}
\alias{get.pathways}
\title{Finds homo sapiens pathways in a patway database}
\usage{
get.pathways(pathwaydatabase = "reactome.db", id = c("symbol",
  "entrez"), min.n.genes = 10)
}
\arguments{
\item{pathwaydatabase}{Currently only "reactome.db" is implemented}

\item{id}{One of "symbol" or "entrez"}

\item{min.n.genes}{Minimal number of genes in a pathway}
}
\value{
A list of pathway ids, athway descriptions, and entrez genes in each pathway
}
\description{
Finds homo sapiens pathways in a patway database
}
\examples{
get.pathways()
}
