% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.pts.R
\name{predict.pts}
\alias{predict.pts}
\title{Method to apply principal curve to new observations}
\usage{
\method{predict}{pts}(obj, newdata, normalize = T)
}
\arguments{
\item{obj}{output from compute.pts}

\item{newdata}{expression data ({genes in pathway} x samples)}

\item{normalize}{Boolean variable indicating whether normalization (substraction by mean, and division by standard deviation) of the rows (genes) should be applied}
}
\value{
A list of pathtracer (pts), pathifier (pds) and p-value
}
\description{
Method to apply principal curve to new observations
}
