% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.pathways.R
\name{get.pathways}
\alias{get.pathways}
\title{Finds homo sapiens pathways in the reactome database}
\usage{
get.pathways(id = c("symbol", "entrez"), min.size = 10)
}
\arguments{
\item{id}{One of symbol or entrez}

\item{min.size}{Minimal number of genes in a pathway}
}
\value{
A list of reactome pathway ids, reactome pathway descriptions, and entrez genes in each pathway
}
\description{
Finds homo sapiens pathways in the reactome database
}
\examples{
get.pathways()
}
