% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi.compute.pts.R
\name{multi.compute.pts}
\alias{multi.compute.pts}
\title{pathtracer}
\usage{
multi.compute.pts(data, reference, ncomp = 4, normalize = T,
  pathwaydatabase = reactome.db, ncores = 1, min.n.genes = 10)
}
\arguments{
\item{data}{Matrix of gene expression data with rows corresponding to genes and columns to samples.}

\item{reference}{Bolean vector indicating which of the columns in \emph{dat} that are reference samples (e.g normal samples).}

\item{ncomp}{Number of principal components.}

\item{normalize}{Boolean. If true, gene expression values are normalized (for each gene, the mean is subtracted and the standard deviation divided upon).}

\item{pathwaydatabase}{Currently, code only for reactome.db is implemented}

\item{min.n.genes}{Minimal number of genes in a pathway required to calculate deregulations scores for the pathway}
}
\value{
res A list of
}
\description{
Computes PathTracer deregulations scores based on gene expressions for a set of pathways. Calls compute.pts
}
\details{
The main PathTracer function. The input is a matrix of gene expressions
for a particular pathway and a set of samples, with rows representing genes, and colums samples.
The function first calculates the principal curve for the first \emph{ncomp} principal components and projects all samples onto the principal curve.
The PathTracer deregulation score (PTS) for each sample is calculated as the Euclidean distance from the sample's projection onto the principal curve to the
reference point (defined as the projection onto the curve for the reference sample with median distance along the curve to the start of the curve). See Nygard et al (2019), for furhter details.
}
\references{
Nygard S, Lingjaerde OC, Caldas C, Hovig E, Borresen-Dahle, Helland Aa, Haakensen V.
"PathTracer: High-sensitivity detection of differential pathway activity in tumours".
}
